Move new code into Test.Kore.Step
